make wrapper class for KCell function so that we can attach sessions to it
