+++ src/kfactory/layout.py
@@ -24,7 +24,7 @@ from pydantic import (
 )
 
 from . import __version__
-from .conf import CheckInstances, config, logger
+from .conf import CheckInstances, config
 from .cross_section import (
     CrossSection,
     CrossSectionModel,
@@ -40,7 +40,7 @@ from .enclosure import (
     LayerEnclosureModel,
     LayerEnclosureSpec,
 )
-from .exceptions import CellNameError, MergeError
+from .exceptions import MergeError
 from .kcell import (
     AnyTKCell,
     BaseKCell,
@@ -64,6 +64,7 @@ from .serialization import (
     _to_hashable,
     get_cell_name,
 )
+from .session_cache import WrappedKCellFunc
 from .settings import Info, KCellSettings, KCellSettingsUnits
 from .typings import K, KCellParams, KCellSpec, MetaData, T
 from .utilities import load_layout_options, save_layout_options
@@ -621,7 +622,7 @@ class KCLayout(
 
         def decorator_autocell(
             f: KCellFunc[KCellParams, AnyTKCell] | KCellFunc[KCellParams, K],
-        ) -> KCellFunc[KCellParams, K]:
+        ) -> WrappedKCellFunc[KCellParams, K]:
             sig = inspect.signature(f)
             output_cell_type_: type[K | ProtoTKCell[Any]]
             if output_type is not None:
@@ -631,256 +632,40 @@ class KCLayout(
             else:
                 output_cell_type_ = self.default_cell_output_type
 
-            output_cell_type = cast("type[K]", output_cell_type_)
+            output_cell_type_ = cast("type[K]", output_cell_type_)
 
             cache_: Cache[_HashedTuple, K] | dict[_HashedTuple, K] = cache or Cache(
                 maxsize=float("inf")
             )
-
-            @functools.wraps(f)
-            def wrapper_autocell(
-                *args: KCellParams.args, **kwargs: KCellParams.kwargs
-            ) -> K:
-                params: dict[str, Any] = {
-                    p.name: p.default for _, p in sig.parameters.items()
-                }
-                param_units: dict[str, str] = {
-                    p.name: p.annotation.__metadata__[0]
-                    for p in sig.parameters.values()
-                    if get_origin(p.annotation) is Annotated
-                }
-                arg_par = list(sig.parameters.items())[: len(args)]
-                for i, (k, _) in enumerate(arg_par):
-                    params[k] = args[i]
-                params.update(kwargs)
-
-                del_parameters: list[str] = []
-
-                for key, value in params.items():
-                    if isinstance(value, dict | list):
-                        params[key] = _to_hashable(value)
-                    elif isinstance(value, kdb.LayerInfo):
-                        params[key] = self.get_info(self.layer(value))
-                    if value is inspect.Parameter.empty:
-                        del_parameters.append(key)
-
-                for param in del_parameters:
-                    params.pop(param, None)
-                    param_units.pop(param, None)
-
-                @cachetools.cached(cache=cache_, lock=RLock())
-                @functools.wraps(f)
-                def wrapped_cell(**params: Any) -> K:
-                    for key, value in params.items():
-                        if isinstance(value, DecoratorDict | DecoratorList):
-                            params[key] = _hashable_to_original(value)
-                    old_future_name: str | None = None
-                    if set_name:
-                        if basename is not None:
-                            name = get_cell_name(basename, **params)
-                        else:
-                            name = get_cell_name(f.__name__, **params)
-                        old_future_name = self.future_cell_name
-                        self.future_cell_name = name
-                        if layout_cache:
-                            if overwrite_existing:
-                                for c in list(self._cells(self.future_cell_name)):
-                                    self[c.cell_index()].delete()
-                            else:
-                                layout_cell = self.layout_cell(self.future_cell_name)
-                                if layout_cell is not None:
-                                    logger.debug(
-                                        "Loading {} from layout cache",
-                                        self.future_cell_name,
-                                    )
-                                    return self.get_cell(
-                                        layout_cell.cell_index(), output_cell_type
-                                    )
-                        logger.debug(f"Constructing {self.future_cell_name}")
-                        name_: str | None = name
-                    else:
-                        name_ = None
-                    cell = f(**params)  # type: ignore[call-arg]
-
-                    logger.debug("Constructed {}", name_ or cell.name)
-
-                    if cell.locked:
-                        # If the cell is locked, it comes from a cache (most likely)
-                        # and should be copied first
-                        cell = cell.dup(new_name=self.future_cell_name)
-                    if overwrite_existing:
-                        for c in list(self._cells(name_ or cell.name)):
-                            if c is not cell.kdb_cell:
-                                self[c.cell_index()].delete()
-                    if set_name and name_:
-                        if debug_names and cell.kcl.layout_cell(name_) is not None:
-                            logger.opt(depth=4).error(
-                                "KCell with name {name} exists already. Duplicate "
-                                "occurrence in module '{module}' at "
-                                "line {lno}",
-                                name=name_,
-                                module=f.__module__,
-                                function_name=f.__name__,
-                                lno=inspect.getsourcelines(f)[1],
-                            )
-                            raise CellNameError(
-                                f"KCell with name {name_} exists already."
-                            )
-
-                        cell.name = name_
-                        self.future_cell_name = old_future_name
-                    if set_settings:
-                        if hasattr(f, "__name__"):
-                            cell.function_name = f.__name__
-                        elif hasattr(f, "func"):
-                            cell.function_name = f.func.__name__
-                        else:
-                            raise ValueError(f"Function {f} has no name.")
-                        cell.basename = basename
-
-                        for param in drop_params:
-                            params.pop(param, None)
-                            param_units.pop(param, None)
-                        cell.settings = KCellSettings(**params)
-                        cell.settings_units = KCellSettingsUnits(**param_units)
-                    if check_ports:
-                        port_names: dict[str | None, int] = defaultdict(int)
-                        for port in cell.ports:
-                            port_names[port.name] += 1
-                        duplicate_names = [
-                            (name, n) for name, n in port_names.items() if n > 1
-                        ]
-                        if duplicate_names:
-                            raise ValueError(
-                                "Found duplicate port names: "
-                                + ", ".join(
-                                    [f"{name}: {n}" for name, n in duplicate_names]
-                                )
-                                + " If this intentional, please pass "
-                                "`check_ports=False` to the @cell decorator"
-                            )
-                    match check_instances:
-                        case CheckInstances.RAISE:
-                            if any(inst.is_complex() for inst in cell.each_inst()):
-                                raise ValueError(
-                                    "Most foundries will not allow off-grid "
-                                    "instances. Please flatten them or add "
-                                    "check_instances=False to the decorator.\n"
-                                    "Cellnames of instances affected by this:"
-                                    + "\n".join(
-                                        inst.cell.name
-                                        for inst in cell.each_inst()
-                                        if inst.is_complex()
-                                    )
-                                )
-                        case CheckInstances.FLATTEN:
-                            if any(inst.is_complex() for inst in cell.each_inst()):
-                                cell.flatten()
-                        case CheckInstances.VINSTANCES:
-                            if any(inst.is_complex() for inst in cell.each_inst()):
-                                complex_insts = [
-                                    inst
-                                    for inst in cell.each_inst()
-                                    if inst.is_complex()
-                                ]
-                                for inst in complex_insts:
-                                    vinst = cell.create_vinst(
-                                        self[inst.cell.cell_index()]
-                                    )
-                                    vinst.trans = inst.dcplx_trans
-                                    inst.delete()
-                        case CheckInstances.IGNORE:
-                            pass
-                    cell.insert_vinsts(recursive=False)
-                    if snap_ports:
-                        for port in cell.to_itype().ports:
-                            if port.base.dcplx_trans:
-                                dup = port.base.dcplx_trans.dup()
-                                dup.disp = self.to_um(
-                                    self.to_dbu(port.base.dcplx_trans.disp)
-                                )
-                                port.dcplx_trans = dup
-                    if add_port_layers:
-                        for port in cell.to_itype().ports:
-                            if port.layer in cell.kcl.netlist_layer_mapping:
-                                if port.base.trans:
-                                    edge = kdb.Edge(
-                                        kdb.Point(0, -port.width // 2),
-                                        kdb.Point(0, port.width // 2),
-                                    )
-                                    cell.shapes(
-                                        cell.kcl.netlist_layer_mapping[port.layer]
-                                    ).insert(port.trans * edge)
-                                    if port.name:
-                                        cell.shapes(
-                                            cell.kcl.netlist_layer_mapping[port.layer]
-                                        ).insert(kdb.Text(port.name, port.trans))
-                                else:
-                                    dwidth = self.to_um(port.width)
-                                    dedge = kdb.DEdge(
-                                        kdb.DPoint(0, -dwidth / 2),
-                                        kdb.DPoint(0, dwidth / 2),
-                                    )
-                                    cell.shapes(
-                                        cell.kcl.netlist_layer_mapping[port.layer]
-                                    ).insert(port.dcplx_trans * dedge)
-                                    if port.name:
-                                        cell.shapes(
-                                            cell.kcl.netlist_layer_mapping[port.layer]
-                                        ).insert(
-                                            kdb.DText(
-                                                port.name,
-                                                port.dcplx_trans.s_trans(),
-                                            )
-                                        )
-                    # post process the cell
-                    for pp in post_process:
-                        pp(cell)  # type: ignore[arg-type]
-                    cell.base.lock()
-                    if cell.kcl != self:
-                        raise ValueError(
-                            "The KCell created must be using the same"
-                            " KCLayout object as the @cell decorator. "
-                            f"{self.name!r} != {cell.kcl.name!r}. Please make sure "
-                            "to use @kcl.cell and only use @cell for cells which "
-                            "are created through kfactory.kcl. To create KCells not"
-                            " in the standard KCLayout, use either "
-                            "custom_kcl.kcell() or KCell(kcl=custom_kcl)."
-                        )
-                    return output_cell_type(base=cell.base)
-
-                with self.thread_lock:
-                    cell_ = wrapped_cell(**params)
-                    if cell_.destroyed():
-                        # If any cell has been destroyed, we should clean up the cache.
-                        # Delete all the KCell entrances in the cache which have
-                        # `destroyed() == True`
-                        deleted_cell_hashes: list[_HashedTuple] = [
-                            _hash_item
-                            for _hash_item, _cell_item in cache_.items()
-                            if _cell_item.destroyed()
-                        ]
-                        for _dch in deleted_cell_hashes:
-                            del cache_[_dch]
-                        cell_ = wrapped_cell(**params)
-
-                    if info is not None:
-                        cell_.info.update(info)
-
-                    return cell_
+            wrapper_autocell = WrappedKCellFunc(
+                kcl=self,
+                f=f,
+                sig=sig,
+                output_type=output_cell_type_,
+                cache=cache_,
+                set_settings=set_settings,
+                set_name=set_name,
+                check_ports=check_ports,
+                check_instances=check_instances,
+                snap_ports=snap_ports,
+                add_port_layers=add_port_layers,
+                basename=basename,
+                drop_params=drop_params,
+                overwrite_existing=overwrite_existing,
+                layout_cache=layout_cache,
+                info=info,
+                post_process=post_process,
+                debug_names=debug_names,
+            )
 
             if register_factory:
                 with self.thread_lock:
-                    if hasattr(f, "__name__"):
-                        function_name = f.__name__
-                    elif hasattr(f, "func"):
-                        function_name = f.func.__name__
-                    else:
+                    if wrapper_autocell.name is None:
                         raise ValueError(f"Function {f} has no name.")
                     if tags:
                         for tag in tags:
                             self.factories.tags[tag].append(wrapper_autocell)
-                    self.factories[basename or function_name] = wrapper_autocell
+                    self.factories[basename or wrapper_autocell.name] = wrapper_autocell
             return wrapper_autocell
 
         return (
